+++ src/quantum/azext_quantum/commands.py
@@ -59,11 +59,11 @@ def transform_jobs(results):
 def transform_users(results):
     def one(result):
         return OrderedDict([
-            ('Principal Id', result.get('principalId')),
-            ('Principal Name', result.get('principalName')),
-            ('Principal Type', result.get('principalType')),
+            ('Name', result.get('displayName')),
+            ('Email', result.get('mail') or result.get('principalName')),
             ('Role', result.get('roleDefinitionName')),
-            ('Scope', result.get('scope'))
+            ('Time Created', result.get('createdOn')),
+            ('Principal Id', result.get('principalId'))
         ])
     return [one(result) for result in results]